requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic-settings==2.6.1
email-validator==2.2.0
pypdf==4.3.1
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
spacy==3.7.5
aiofiles==24.1.0
websockets==14.1
